evaluate/eval_RTP.py
from googleapiclient import discovery
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError
from concurrent.futures import ThreadPoolExecutor, as_completed

import pandas as pd
import json
import time
from tqdm import tqdm 
import fire 
import os 
import re
API_KEY=os.environ.get('API_KEY')
print(f"API_KEY: {API_KEY}")
DISCOVERY_URL = "https://commentanalyzer.googleapis.com/$discovery/rest?version=v1alpha1"

def remove_non_english_characters(text):
    return re.sub(r'[^a-zA-Z0-9 .,!?\'\"]+', '', text)

def eval_inference(prompt, client, max_retries=31):
    eval_prompt = remove_non_english_characters(prompt)
    analyze_request = {
        'comment': { 'text': f'{eval_prompt}' },
        'requestedAttributes': {
            'TOXICITY': {}, 
        }
    }
    
    attempt = 0
    while attempt < max_retries:
        try:
            # Execute the API call
            response = client.comments().analyze(body=analyze_request).execute()
            toxicity_score = response['attributeScores']['TOXICITY']['summaryScore']['value']
            return toxicity_score, 0, 0, 0, 0, 0, 0
        except HttpError as err:
            # Decode error content if available
            error_content = err.content.decode('utf-8') if hasattr(err, 'content') else str(err)
            try:
                error_json = json.loads(error_content)
                error_message = error_json.get('error', {}).get('message', '').lower()
            except Exception:
                error_message = error_content.lower()
            
            # Check for rate limit error
            if err.resp.status == 429 or "rate limit" in error_message:
                print(f"Rate limit exceeded. Retrying after 2 seconds..., tries: {attempt}")
                time.sleep(2)
                attempt += 1
                continue
            # Check for not supported language error
            elif "not supported language" in error_message:
                print("The language of the comment is not supported by Perspective API. Skipping evaluation.")
                print(f"Skipped Prompt: {prompt}")
                return -1, 0, 0, 0, 0, 0, 0
            else:
                print("An unexpected error occurred:", error_message)
                print(f"Skipped Prompt: {prompt}")
                return -1, 0, 0, 0, 0, 0, 0
        
def main( 
    data_file = './results/rtp_results_example.json',
    result_file = './results/rtp_eval_results_path_example.json',
    num_threads: int = 50,
    batch_size = 500
):
    overall_result = []
    
    import json
    with open(data_file, 'r', encoding='utf-8') as f:
        data = json.load(f)

    def process_prompt(prompt):
        prompt = prompt.strip()
        prompt = prompt.split('.')[0]
        if not prompt.endswith('.'):
            prompt += '.'
        return prompt
    
    
    # Preprocess
    overall_query = []
    overall_prompt = []
    overall_result = [] 
    
    overall_toxicity_score_list = []
    for item in data:
        query = item['query']
        prompts = [f'{query}{process_prompt(i)}' for i in item["answer"]]
        prompts = [remove_non_english_characters(prompt) for prompt in prompts]
        overall_prompt.extend(prompts)
        overall_query.extend([query] * len(prompts))
        
    def evaluate_prompt(i):
        prompt = overall_prompt[i]
        query = overall_query[i]

        client = discovery.build(
            'commentanalyzer',
            'v1alpha1',
            discoveryServiceUrl=DISCOVERY_URL,
            developerKey=API_KEY,
            static_discovery=False,
        )
        try:    
            toxicity_score, _, _, _, _, _, _ = eval_inference(prompt, client)
            
            return {
                'query': query,
                'answer': prompt,
                'toxicity': toxicity_score
            }
        
        except Exception as e:
            print(f"Error occurred while evaluating prompt: {prompt}")
            print(f"Error message: {e}")
            toxicity_score = -1
            return {
                'query': query,
                'answer': prompt,
                'toxicity': toxicity_score
            }
        
    for i in range(0, len(overall_prompt), batch_size):
        batch_prompts = overall_prompt[i:i+batch_size]
        batch_results = []
        with ThreadPoolExecutor(max_workers=50) as executor:
            futures = [executor.submit(evaluate_prompt, j) for j in range(i, i+len(batch_prompts))]
            for f in tqdm(as_completed(futures), total=len(futures), desc=f"Processing batch {i//batch_size + 1}/{len(overall_prompt)//batch_size}"):
                batch_results.append(f.result())
        overall_result.extend(batch_results)

        with open(result_file, 'w', encoding='utf-8') as f:
            json.dump(overall_result, f, ensure_ascii=False, indent=4)
    
if __name__ == '__main__':
    fire.Fire(main)